Add support to the REPL to let users type in both statements and expressions
This wasn't entirely easy to get done in a clean way. I chose the
approach of trying to parse it as a set of statements first, then if it
fails, as an expression, but I also had to change the error handling to
not output errors if the first parsing fails... hence, the added quirks
in this commit.